+++ tests/testthat/helper-traduire.R
@@ -4,6 +4,8 @@ hintr_translator <- function() {
 
 with_hintr_language <- function(language, expr) {
   reset <- traduire::translator_set_language(language, package = "hintr")
-  on.exit(reset())
+  naomi_reset <- traduire::translator_set_language(language, package = "naomi")
+  on.exit(reset(), add = TRUE)
+  on.exit(naomi_reset(), add = TRUE)
   force(expr)
 }